+++ microservice-producto-categoria/src/main/java/com/microservice/producto_categoria/infraestructure/abstract_services/IProductoService.java
@@ -4,5 +4,5 @@ import com.microservice.producto_categoria.api.domain.request.ProductoRequest;
 import com.microservice.producto_categoria.api.domain.responses.ProductoResponse;
 import com.microservice.producto_categoria.domain.entities.Producto;
 
-public interface IProductoService extends CrudService<ProductoRequest, ProductoResponse, Long> {
+public interface IProductoService extends CrudService<ProductoRequest, ProductoResponse, Integer> {
 }
